Refactor auth and error handling
Instead of using a different handler type for errors, the auth package
now instead provides the necessary functions to check if a user has
sufficient priveleges in any handler.

The apperr.Error type was expanded to include additional information
about where the error came from.

A new function, apperr.HTTPError, was added to increase allow for a more
ergonomic error API by allowing the use of the unwrapped error
interface.